search_works_adapter_works
